Seperated out all components to their own headers
