+++ Sol/Source/Sol/Scene/Components/NameComp.h
@@ -1,0 +1,19 @@
+#pragma once
+
+#include "Sol/SolDefines.h"
+
+namespace Sol 
+{
+	struct NameComp
+	{
+		NameComp() = default;
+		NameComp(const NameComp&) = default;
+		NameComp(const std::string& name) :m_Name(name) {}
+		~NameComp() = default;
+
+		operator std::string& () { return m_Name; }
+		operator const std::string& () const { return m_Name; }
+	private:
+		std::string m_Name;
+	};
+ }
